+++ examples/functions/count/function.tf
@@ -1,0 +1,3 @@
+# Return the substring count of a string:
+provider::stdlib::count("pizza", "z")
+# result => 2